Move PasswordToggleElement out of index
**Why**: So we can have multiple exports
